requirements.txt
# AI and Machine Learning
tensorflow==2.9.1
pytorch==1.12.0
scikit-learn==1.1.2
pandas==1.4.3
numpy==1.23.1
matplotlib==3.5.2

# Blockchain Integration
web3==5.30.0
solana==0.27.2
eth-brownie==1.19.2

# API and Web
fastapi==0.79.0
uvicorn==0.18.2
requests==2.28.1
websockets==10.3

# Database
pymongo==4.2.0
motor==3.0.0

# Utils
python-dotenv==0.20.0
pydantic==1.9.1
pytest==7.1.2
black==22.6.0
isort==5.10.1 